+++ TPServer/PacketProcessor.h
@@ -9,7 +9,7 @@ class Session;
 class PacketProcessor : public TSingleton<PacketProcessor>
 {
 public:
-	void Process(Session* const owner, char* const buffer, const DWORD bytesTransferred);	
+	void Process(Session* const owner, char* const buffer, const size_t bytesTransferred);
 	void SendPacket(const Packet& packet);
 	void SendPacket(const Packet& packet, const Session* const session);
 	void SendPacket(const Packet& packet, const SOCKET& clntSock);
